[Travis] Allow to keep xdebug enabled

+++ bin/.travis/prepare_system.sh
@@ -9,7 +9,9 @@
 ##       https://github.com/facebook/hiphop-php/commit/0b2dfdf4492eb06a125b068e939d092ec0588e5c
 
 # Disable xdebug to speed things up
-phpenv config-rm xdebug.ini
+if [ "$DEBUGGING" == "" -a "$TRAVIS_PHP_VERSION" != "" -a "$TRAVIS_PHP_VERSION" != "hhvm" ] ; then
+    phpenv config-rm xdebug.ini ;
+fi
 
 # Install needed packages
 echo "> Installing needed packages";
